requirements.txt
rich
dnspython
requests
python-whois
aiohttp
shodan
tabulate
python-nmap
python-whois
